+++ src/app/locales/en/components/storage-import-update-form.js
@@ -4,4 +4,8 @@ import storageUpdate from './-space/storage-update';
 export default {
   storageImport,
   storageUpdate,
+  importConfiguration: 'Import configuration',
+  updateConfiguration: 'Update configuration',
+  submitNew: 'Start synchronization',
+  submitModify: 'Save configuration',
 };