+++ src/app/mixins/components/space-item-sync-stats.js
@@ -3,19 +3,19 @@
  *
  * @module mixins/space-item-sync-stats
  * @author Jakub Liput
- * @copyright (C) 2017 ACK CYFRONET AGH
+ * @copyright (C) 2017-2019 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
 import Mixin from '@ember/object/mixin';
 import { inject as service } from '@ember/service';
-import { computed, observer, set, get } from '@ember/object';
+import { computed, observer, get } from '@ember/object';
 import { readOnly, equal, not } from '@ember/object/computed';
 import { isEmpty } from '@ember/utils';
 import _ from 'lodash';
 import moment from 'moment';
 import Looper from 'onedata-gui-common/utils/looper';
-import safeMethodExecution from 'onedata-gui-common/utils/safe-method-execution';
+import safeExec from 'onedata-gui-common/utils/safe-method-execution';
 
 /**
  * How often watchers should sample for stats for given metrics [ms]
@@ -150,7 +150,7 @@ export default Mixin.create({
     // interval of this Looper will be set in reconfigureSyncWatchers observer
     let _syncChartStatsWatcher = Looper.create({ immediate: true });
     _syncChartStatsWatcher.on('tick', () =>
-      safeMethodExecution(this, 'fetchAllSyncStats')
+      safeExec(this, 'fetchAllSyncStats')
     );
 
     let _syncStatusWatcher = Looper.create({
@@ -158,7 +158,7 @@ export default Mixin.create({
       interval: this.get('syncStatusRefreshTime'),
     });
     _syncStatusWatcher.on('tick', () =>
-      safeMethodExecution(this, 'checkSyncStatusUpdate')
+      safeExec(this, 'checkSyncStatusUpdate')
     );
     this.checkSyncStatusUpdate();
 
@@ -212,11 +212,13 @@ export default Mixin.create({
     let currentSyncStats = this.get('_syncStats');
     if (currentSyncStats != null) {
       // we already got some syncStats so update only statuses
-      set(currentSyncStats, 'importStatus', get(newSyncStats, 'importStatus'));
-      set(currentSyncStats, 'updateStatus', get(newSyncStats, 'updateStatus'));
+      safeExec(this, 'setProperties', {
+        importStatus: get(newSyncStats, 'importStatus'),
+        updateStatus: get(newSyncStats, 'updateStatus'),
+      });
     } else {
       // first syncStats update
-      this.set('_syncStats', newSyncStats);
+      safeExec(this, 'set', '_syncStats', newSyncStats);
     }
   },
 
@@ -236,9 +238,9 @@ export default Mixin.create({
   },
 
   fetchAllSyncStats() {
-    let syncInterval = this.get('syncInterval');
+    const syncInterval = this.get('syncInterval');
 
-    let syncStatsPromise =
+    const syncStatsPromise =
       this.get('spaceManager').getSyncAllStats(
         this.get('space.id'),
         syncInterval
@@ -251,7 +253,7 @@ export default Mixin.create({
         this.get('space.importEnabled') &&
         get(newSyncStats, 'importStatus') === 'done') {
 
-        safeMethodExecution(this, 'set', 'statsFrozen', true);
+        safeExec(this, 'set', 'statsFrozen', true);
       }
 
       this.setProperties({
@@ -277,24 +279,22 @@ export default Mixin.create({
 
   /**
    * Is sync tab currently opened
-   * NOTE: activeTabId is provided by `space-tabs` mixin
+   * NOTE: selectedTab is provided by `space-tabs` mixin
    * @type {Ember.ComputedProperty<boolean>}
    */
-  syncTabActive: computed('activeTabId', function () {
-    const activeTabId = this.get('activeTabId');
-    return activeTabId && /^tab-sync-/.test(activeTabId);
+  syncTabActive: computed('selectedTab', function () {
+    const selectedTab = this.get('selectedTab');
+    return selectedTab && /^tab-sync-/.test(selectedTab);
   }),
 
   reconfigureSyncWatchers: observer(
-    '_isActive',
     '_importActive',
     'syncInterval',
     '_syncChartStatsWatcher',
     'statsFrozen',
     'syncTabActive',
     function () {
-      let {
-        _isActive,
+      const {
         _importActive,
         syncInterval,
         _syncChartStatsWatcher,
@@ -303,7 +303,6 @@ export default Mixin.create({
         syncStatusRefreshTime,
         syncTabActive,
       } = this.getProperties(
-        '_isActive',
         '_importActive',
         'syncInterval',
         '_syncChartStatsWatcher',
@@ -317,7 +316,7 @@ export default Mixin.create({
         _syncStatusWatcher.set('interval', syncStatusRefreshTime);
       }
 
-      if (syncTabActive && _importActive && _isActive && !statsFrozen) {
+      if (syncTabActive && _importActive && !statsFrozen) {
         _syncChartStatsWatcher.set('interval', WATCHER_INTERVAL[syncInterval]);
       } else {
         _syncChartStatsWatcher.stop();