+++ src/lib/ember-onedata-onepanel-server/addon/services/onepanel-server-mock.js
@@ -5,7 +5,7 @@
  *
  * @module services/onepanel-server-mock
  * @author Jakub Liput
- * @copyright (C) 2017-2018 ACK CYFRONET AGH
+ * @copyright (C) 2017-2019 ACK CYFRONET AGH
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
@@ -104,11 +104,11 @@ function _genSupportingProviders() {
 
 function _genAutoCleaningConfiguration() {
   return {
-    lowerFileSizeLimit: {
+    minFileSize: {
       enabled: true,
       value: 10000,
     },
-    upperFileSizeLimit: {
+    maxFileSize: {
       enabled: true,
       value: 10000000,
     },
@@ -433,7 +433,7 @@ export default OnepanelServerBase.extend(
 
           if (mockStep >= STEP.PROVIDER_DONE) {
             const spaces = this.get('__spaces');
-            const spacesFilesPopularity = this.get('__spacesFilesPopularity');
+            const spacesFilePopularity = this.get('__spacesFilePopularity');
             const spacesAutoCleaning = this.get('__spacesAutoCleaning');
             spaces.push({
               id: 'space1_verylongid',
@@ -449,10 +449,13 @@ export default OnepanelServerBase.extend(
               },
               supportingProviders: _genSupportingProviders(),
             });
-            spacesFilesPopularity.push({
+            spacesFilePopularity.push({
               id: 'space1_verylongid',
               enabled: true,
-              restUrl: 'https://example.com',
+              exampleQuery: 'curl https://example.com',
+              lastOpenHourWeight: 2,
+              avgOpenCountPerDayWeight: 3,
+              maxAvgOpenCountPerDay: 4,
             });
             spacesAutoCleaning.push({
               id: 'space1_verylongid',
@@ -866,19 +869,26 @@ export default OnepanelServerBase.extend(
       };
     },
 
-    _req_oneprovider_configureFilesPopularity() {
+    _req_oneprovider_configureFilePopularity() {
       return {
-        // data: { enabled, threshold, target, rules }
+        // data: enabled, [lastOpenHourWeight], [avgOpenCountPerDayWeight], [maxAvgOpenCountPerDay]
         success: (id, data) => {
-          const spacesFilesPopularity = this.get('__spacesFilesPopularity');
-          let configuration = spacesFilesPopularity.find(s => s.id === id);
+          const spacesFilePopularity = this.get('__spacesFilePopularity');
+          let configuration = spacesFilePopularity.find(s => s.id === id);
           if (!configuration) {
             configuration = { id };
-            spacesFilesPopularity.push(configuration);
+            spacesFilePopularity.push(configuration);
           }
           const popEnabled = get(data, 'enabled');
           if (popEnabled === true) {
-            set(data, 'restUrl', 'https://example.com/api/2');
+            setProperties(
+              configuration,
+              Object.assign({
+                  exampleQuery: 'curl https://example.com/api/2',
+                },
+                data
+              )
+            );
           } else if (popEnabled === false) {
             let autoCleaningConfiguration = this.get('__spacesAutoCleaning').find(s =>
               s.id === id
@@ -890,23 +900,23 @@ export default OnepanelServerBase.extend(
           emberObjectMerge(configuration, data);
         },
         statusCode: (id) => {
-          const spacesFilesPopularity = this.get('__spacesFilesPopularity');
-          let configuration = spacesFilesPopularity.find(s => s.id === id);
+          const spacesFilePopularity = this.get('__spacesFilePopularity');
+          let configuration = spacesFilePopularity.find(s => s.id === id);
           return configuration ? 204 : 404;
         },
       };
     },
 
-    _req_oneprovider_getFilesPopularityConfiguration() {
+    _req_oneprovider_getFilePopularityConfiguration() {
       return {
         success: (id) => {
-          const __spacesFilesPopularity = this.get('__spacesFilesPopularity');
-          const configuration = __spacesFilesPopularity.find(s => s.id === id);
+          const __spacesFilePopularity = this.get('__spacesFilePopularity');
+          const configuration = __spacesFilePopularity.find(s => s.id === id);
           return _.cloneDeep(configuration);
         },
         statusCode: (id) => {
-          const __spacesFilesPopularity = this.get('__spacesFilesPopularity');
-          const configuration = __spacesFilesPopularity.find(s => s.id === id);
+          const __spacesFilePopularity = this.get('__spacesFilePopularity');
+          const configuration = __spacesFilePopularity.find(s => s.id === id);
           return configuration ? 204 : 404;
         },
       };
@@ -983,7 +993,15 @@ export default OnepanelServerBase.extend(
 
     _req_oneprovider_getProviderSpaceAutoCleaningReports: computed(function () {
       return {
-        success: (id, startedAt) => this._getReportsCollection(id, startedAt),
+        success: (spaceId, { index, limit, offset }) =>
+          this._getReportIds(spaceId, index, limit, offset),
+        statusCode: () => 200,
+      };
+    }),
+
+    _req_oneprovider_getProviderSpaceAutoCleaningReport: computed(function () {
+      return {
+        success: (spaceId, reportId) => this._getReport(spaceId, reportId),
         statusCode: () => 200,
       };
     }),
@@ -1320,7 +1338,7 @@ export default OnepanelServerBase.extend(
 
     __spaces: A([]),
 
-    __spacesFilesPopularity: A([]),
+    __spacesFilePopularity: A([]),
 
     __spacesAutoCleaning: A([]),
   });